+++ Projects/Modules/KeychainModule/Sources/DefaultKeychain.swift
@@ -1,0 +1,42 @@
+import Foundation
+
+public struct DefaultKeychain: Keychain {
+    private let service: String = Bundle.main.bundleIdentifier ?? ""
+
+    public func save(type: KeychainType, value: String) {
+        let query: NSDictionary = [
+            kSecClass: kSecClassGenericPassword,
+            kSecAttrService: service,
+            kSecAttrAccount: type.rawValue,
+            kSecValueData: value.data(using: .utf8, allowLossyConversion: false) ?? .init()
+        ]
+        SecItemDelete(query)
+        SecItemAdd(query, nil)
+    }
+
+    public func load(type: KeychainType) -> String {
+        let query: NSDictionary = [
+            kSecClass: kSecClassGenericPassword,
+            kSecAttrService: service,
+            kSecAttrAccount: type.rawValue,
+            kSecReturnData: kCFBooleanTrue!,
+            kSecMatchLimit: kSecMatchLimitOne
+        ]
+        var dataTypeRef: AnyObject?
+        let status = withUnsafeMutablePointer(to: &dataTypeRef) { SecItemCopyMatching(query, UnsafeMutablePointer($0)) }
+        if status == errSecSuccess {
+            guard let data = dataTypeRef as? Data else { return "" }
+            return String(data: data, encoding: .utf8) ?? ""
+        }
+        return ""
+    }
+
+    public func delete(type: KeychainType) {
+        let query: NSDictionary = [
+            kSecClass: kSecClassGenericPassword,
+            kSecAttrService: service,
+            kSecAttrAccount: type.rawValue
+        ]
+        SecItemDelete(query)
+    }
+}
