+++ Projects/Modules/KeychainModule/Sources/Keychain.swift
@@ -6,46 +6,8 @@ public enum KeychainType: String {
     case expiredAt = "EXPIRED-AT"
 }
 
-public struct Keychain {
-    private let service: String
-    public init(service: String = Bundle.main.bundleIdentifier ?? "") {
-        self.service = service
-    }
-
-    public func save(type: KeychainType, value: String) {
-        let query: NSDictionary = [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrService: service,
-            kSecAttrAccount: type.rawValue,
-            kSecValueData: value.data(using: .utf8, allowLossyConversion: false) ?? .init()
-        ]
-        SecItemDelete(query)
-        SecItemAdd(query, nil)
-    }
-
-    public func load(type: KeychainType) -> String {
-        let query: NSDictionary = [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrService: service,
-            kSecAttrAccount: type.rawValue,
-            kSecReturnData: kCFBooleanTrue!,
-            kSecMatchLimit: kSecMatchLimitOne
-        ]
-        var dataTypeRef: AnyObject?
-        let status = withUnsafeMutablePointer(to: &dataTypeRef) { SecItemCopyMatching(query, UnsafeMutablePointer($0)) }
-        if status == errSecSuccess {
-            guard let data = dataTypeRef as? Data else { return "" }
-            return String(data: data, encoding: .utf8) ?? ""
-        }
-        return ""
-    }
-
-    public func delete(type: KeychainType) {
-        let query: NSDictionary = [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrService: service,
-            kSecAttrAccount: type.rawValue
-        ]
-        SecItemDelete(query)
-    }
+public protocol Keychain {
+    func save(type: KeychainType, value: String)
+    func load(type: KeychainType) -> String
+    func delete(type: KeychainType)
 }
